Fix relocatable builds on the moar backend
For relocatability to work the config variable `static_rakudo_home`
needs to be empty. It defaulted to the build time rakudo home path
globally. Later on a MoarVM specific config variable with the same name
was set to empty assuming it would overwrite the global variable.
But that's not the case. In backend specific templates the empty
variable would be used, but in backend unspecific templates one still
accessed the global variable. This caused the backend-unspecific
main-version.in template to still pick up the wrong global
static_rakudo_home variable.

Fix this by setting the static_rakudo_home variable (which is
intrinsically backend unspecific) correctly at the global level. No
moar specific variable with the same name is created anymore.

Fixes #3669