Remove return constraint from identity methods
These would be the methods that always return "self".  By removing
this constraint, these methods became about 1.5x as fast.  Now, they
didn't do much to begin with, so in absolute numbers this probably
won't matter much.  But in some cases, such as Str.gist, this may
actually be noticeable in some applications that do a lot of "say"
on strings.