Normalizing IO::Handle.say/put/print/note, part 2
The &say sub now performs all of the .gist + newline logic and then
dispatches to .print.  The advantage of this is that we do not need
any additional logic / checks for modules that pretend to be $*OUT.

This also removes the IO::Handle.say(Str:D) candidate: even though a
Str may be passed to IO::Handle.say, it should *still* call the .gist
method on it, as it can have a non noop .gist method mixed in (roast
actually tests for this in the &sub case).

This commit also creates a single proto say for all of Rakudo.

This commit also adds Mu.say candidates to take care of those classes
that pretend to do the $*OUT API and which do not provide their own
.say method.

Finally, this adds two Rakudo::Internals methods for quickly generating
a native str array from a given list, which each element .gist or .Str'd.