Improve error message for too many directives in routine `sprintf`
Fix GH issue #3682

The problem is the erroneous use of a '$' symbol in the sprintf format
statement does not give a hint of that possibility in the original
exception code. That, coupled with incomplete sprintf documentation,
may mislead the user who makes that mistake.

The fix is to mention the possibility of the misused '$' symbol when
the args provided are less than the detected directives in the format
string.

+ modify and improve file rakudo/src/core.c/Exception.pm6
+ expand and improve core test file rakudo/t/05-messages/02-errors.t