Add a testing configuration to Azure Pipelines
This configuration tests a wide array of build combinations:
- Windows, MacOS and Linux
- MoarVM and JVM
- NQP master and tag
- MoarVM master and tag
- Relocatable

This is about everything that AppVeyor, CircleCI and Travis do in
combination. So the mid-term plan is to get rid of the other three
CI services. This should save a good bunch of resources.

Also the build artifacts of the relocatable build are published.
These are only useful for testing as they are not built the same
way as actual releases are and thus inhibit some limitations.
They do not contain the auxiliary files (README and similar) and the
Linux build is created on a recent Ubuntu instead of an ancient CentOS
which results in a binary that will only run on recent Linuxes.