Fix CI failures when commits are added fast
With a fetch depth of 1 azure pipelines will always clone the master branch
with only the top most commit and then try to check out the commit that
is to be tested. This fails when commits are added quickly, because then
test jobs that weren't fast enough to start will check clone only the newly
added commit and then checking out the commit to test fails because that's
not part of the cloned repository.