Fix build artifact publishing
Artifacts names must be unique. When trying to publish an artifact with a
name that already exists an "ArtifactExistsException" is thrown and the
job fails. To solve this we don't explicitly specify the artifact name
which should then default to a name autogenerated based on the job name.
See https://github.com/MicrosoftDocs/azure-devops-docs/issues/6765#issue-541875716