Partially fix building the JVM in AzureCI
Installation still fails with a different error.

Windows has a maximum PATH variable length of 2048 (depending on
how it's accessed). The length of PATH in AzureCI is already
really tight. When building the JVM backend we run into the limit
when we add Java and the MS BuildTools to the path.
To work around this, we remove a bunch of stuff we won't need
from PATH.