Use hllbool(istrue) for so() and not() in the end
Turns out that *is* the fastest way to get a Bool for an expression.
The ternary case appears to call "identity" according to the profile,
whereas the hllbool case does not.  Although this may be an artefact
of the profile, it does show that it is doing more for the ternary
case, then it is for the hllbool case.