Add RAKU_REPL_OUTPUT_METHOD environment variable
When set, it should contain the name of the method with which the
value of an expression should be output.  The default is "gist",
which is the current state.  Expected alternate values are "raku"
and "Str".  But of course you can go crazy in this, and some of
you probably will.

This in response to the discussion in:

   https://github.com/Raku/problem-solving/issues/194

Please revert if consensus is that this is a bad idea.  Personally
I would probably use RAKU_REPL_OUTPUT_METHOD=raku by default, but
that is just me :-)