Internally rename IO::Path's $!abspath to $!os-path
It took me quite some time to realize the subtle bug in my head that
assumed that $!abspath would always be the same as $!path if the $!path
was specified as an absolute path.  $!abspath is the result of .absolute,
which calls $!SPEC.rel2abs($!path), which may result in something
entirely different, such as:

    dd IO::Path::Win32.new("/foo").absolute; # "\\foo"

By changing the name, hopefully future maintainers won't fall into this
trap.  This also implies that binding $!abspath to $!path is basically
an optimization that will break code in some cases.  But that's for
another commit.