Make WhateverCode.ACCEPTS about 2x as fast
Since a WhateverCode must always have at least on parameter, and
ACCEPTS only accepts one parameter as well, there is no need to
check any signature for WhateverCode, so give it its own candidate
and directly use an nqp:: call to execute.  This halves the Exclusive
Time for ACCEPTS in a profile.