Add 'auth-fmt' compiler arg for option '--doc'
Addresses Rakudo issue GH #3701.

This commit adds the capability for the user to affect parsing of Pod
in leading declarator blocks.

The standard behavior is to extract such Pod as a single line of
normalized text. A previous commit added the capability to change such
parsing to respect the original formatting by saving the author's
newlines and other spaces, but it had to be triggered by the existence
of a special Rakudo environment variable:

  RAKUDO_POD_DECL_BLOCK_USER_FORMAT

With this commit, the user can now select the same behavior by adding
the word 'auth-fmt' after the command line option '--doc='.  The new
word can be added alone or added to the existing argument choices of
'Text' or 'HTML'. The two-word version must have the words separated
by a comma (with no spaces) and in any order.

The compiler has been modified so the '--help' option explains the new
capability.

Files modified:

+ src/Perl6/Compiler.nqp
+ src/Perl6/Grammar.nqp
+ src/main.nqp