Make sure .cloned-with-path performs null check
This was not picked up by spectesting so far, but it did show in the
IO::Path.child tests.  This makes all of these .cloned-with-path
optimizations a little bit slower, e.g. the .child one from 1.4x as
fast to 1.25x as fast.

Should probably need to make a separate function without the null-check
for uses such as dir(), where we can assume that directory entries do
not contain nulls.

This should make CI happy again.