Simplify R:I.FILETEST_x methods
All of these methods return a native uint, and will throw when a path does
not exist in some cases (the ones using nqp::stat) or just return 0.  It
does not make sense to check for weird values, when such values indicate a
deeper issue.

Also make the conversion to native str in the signature, so that places
that provide a native str would not need to HLLize just to be able to
call the method.

This should probably be noticeable in some file-heavy applications.