Add '--doc-fmt=preserve' compiler option
Addresses Rakudo issue GH #3701.

This commit adds the capability for the user to affect parsing of Pod
in leading declarator blocks.

The standard behavior is to extract such Pod as a single line of
normalized text. A previous commit added the capability to change such
parsing to respect the original formatting by saving the author's
newlines and other spaces, but it had to be triggered by the existence
of a special Rakudo environment variable:

  RAKUDO_POD_DECL_BLOCK_USER_FORMAT

With this commit, the user can now select the same behavior by adding
the new option '--doc-fmt=preserve' option along with other
'--doc' options.

The compiler has been modified so the '--help' option explains the new
option.

As an example consider this short Raku program in file 't.raku':

    #| line 1: foo
    #| line 2: bar
    sub baz() {}
    #= line 3: baz
    #= line 4: boo

Run using current usage:

    $ raku --doc t.raku
    sub baz()
    line 1: foo line 2: bar
    line 3: baz line 4: boo

Run with the new option:

    $ raku --doc --doc-fmt=preserve t.raku
    sub baz()
    line 1: foo
    line 2: bar
    line 3: baz line 4: boo

Notice the trailing declarator block hasn't been fixed yet. That is a
separate issue.

Files modified:

+ src/Perl6/Compiler.nqp
+ src/Perl6/Grammar.nqp
+ src/main.nqp