Add support for :secure in IO::Path.child
- shamelessly transmogrified the code from IO::Path::ChildSecure
- added test-file to "make spectest"

So basically, to get the IO::Path::ChildSecure semantics, you will need
to pass the ":secure" named parameter to .child as well.  The IO::Path.add
method is *not* affected by this.

Making :secure default to True, made all sorts of spectests fail, so
that does not appear to be an option.  Whether or not .add also should
support an :secure named parameter, is left as an exercise in bikeshedding.