Fix JVM backend (and MoarVM as well)
Turns out nqp::p6argvmarray can only be used in the same *frame* as the
routine it wants the arguments from.  This apparently was covered up on
MoarVM by the static optimizer, but not so on the JVM which uncovered
the issue.

bartolin++ for following up and finding the cause of breakage on JVM