Implement set equality operators
- (==)  returns whether operands are the same using QuantHash semantics
- ≡     unicode version of (==)
- ≢     returns whether operrands are *NOT* the same using QuantHash semantics

Although "a (==) b" can be considered short for "a.Set (==) b.Set", it will
coerce its operands to the appropriate level of QuantHash as appropriate as
all of the other set operators.  And short-circuit as soon as it becomes
clear that the operands are *not* equal.

This allows for things like:

    if <a b c> (==) %options {
        # %options contains just "a" "b" and "c" as keys, and each of
        # these keys have a truthy value
    }

    my @a = 1,2,3;
    my @b = 3,1,2;
    if @a (==) @b {
        # all unique values in @a are found in @b and no more
    }