Don't check whether to actually close when calling .close
61046d7695615d67 unified some logic regarding closing of file handles.
Unfortunately, that broke the semantics regarding .lock/.unlock,
specifically wrt closing a locked file.  Fixed by undoing the unification
of that logoc: only DESTROY will now not close an open handle if it was
told not to do so.  An explicit call to .close will *always* close the
file handle.  Spotted by ugexe++.  Fixes #3742