Bump NQP to get new "-" behaviour
When you start Raku with "-" as a single positional parameter, it
will now check if STDIN is connected to a terminal.  If it is, then
the REPL will be started, rather than appearing to freeze waiting
for the user to enter something and issue an EOF.  If STDIN is *not*
connected to a terminal, then Raku will read from STDIN and process
that as the source of a program.