Fix IO::Path.slurp for special files
If a file claimed it only had 0 bytes, then no attempt would be made
to actually read from that file.  This semantic however broke the
Linux::Cpuinfo module in the ecosystem, because that is trying to
read the /proc/cpuinfo special file that does not claim to consist
of any data.

Now, if a file claims to have 0 bytes, then it will be read in 1M
increments until it produces less than that.  This should fix
Linux::Cpuinfo, but alas, cannot verify as committer is on MacOS.