Fix code objects created BEGIN time EVAL getting lost in precompilation
This is a follow up to the actual fixes in NQP to add frames created by a
nested compiler (e.g. BEGIN time EVAL, traits or role bodies) to the outer
compilation unit, so they make it into a precomp file.

This also replaces the need for backend specific mast_frames argument with
more generic handling by the backend itself.