RakuAST handling of chaining operators
Whether something is chaining is a property known by the parser, thus at
the RakuAST level it seems reasonable to have that modeled as a separate
element type. We can't just have a list of chaining ops in the AST, as
we might end up with user-defined chaining ops also. The alternatives
are to pass properties into the RakuAST nodes (but which ones?) or to
interrogate metadata on the op itself (but that's rather painful for
bootstrapping).