Make sure push-all doesn't push when exhausted
Fixes R#3758.  The sample code would call "pull-one" twice, and then
call "push-all" to fill the array.  A next call to "pull-one" would
have returned an IterationEnd.  The push-all did not account for the
situation where the iterator was basically exhausted, but had not
seen an IterationEnd yet.  I wonder how many other iterators suffer
from this shortcoming.  But that's for another day.