Move all of INTERPOLATE and friends into a separate file
The whole INTERPOLATE business is a entirely different can of worms
that will need some TLC in the not too distant future.  Split it off
into a separate file now so that will be easier to re-implement the
other parts of the Match class.

No functional changes whatsoever.  The git history has been preserved.