Give Grammar its own .new
One should only call .new on a grammar if it has attributes of its
own that need initializing.  This is now (also) performed by
Match.new.  But that is *also* used to create the cursor that will
be used in matching.

By giving Grammar its own new, we're separating Grammar initialization
from cursor initialization.  Don't expect too much of a performance
gain yet, but this will make refactoring Match a lot easier, and allow
that refactor to be much more efficient.