Completely refactor the Match class
The old implementation made Match a subclass of Capture.  This had
the advantage that it could use the Positional and Associative
interface of Capture.  This has the *dis*advantage of needing to
set up a lot of data-structures *after* a Match object is returned
from NQP.  Even if these data structures are not needed.

The aim of this refactor is to create a class that can take the Match
objects returned by NQP to be used *without* needing to set up any
additional data structures.  The data structure building method
Match.MATCH is now an empty shell: calls to it are still to be removed.

This is now test-clean, but not all places use the new situation to
its full advantage.

A Match object contains other Match objects if it has any positional
or named captures.  These captures may be slightly more expensive to
get because of the lack of extra data structures.  However, the leaf
Match objects, will not need any additional processing at all, and
those ones are the most occurring Match objects.