Fix reporting of unhandled exceptions caused by NQP vs. Raku arrays
The code generated for handling END phasers expected @END_PHASERS to have an
unshift method while the code for handling exceptions expected it to be a
VMArray object. Both expectations can only be met by using an NQPArray which
is hard to create from Raku code. So instead, use nqp::unshift instead of the
method which is faster anyway.