Make sure that Any:U.head/tail work
This seems to be implied by Any:U.skip working, and a recent change
broke a spectest about that.  Now, this breaks tests for Supply:U.head
and Supply:U.tail, but given the discussion about Any:U.skip in R#1384,
it looks to me that those tests are in error.