Make <a b c d>.permutations between 23 and 85x faster
When running that code 2000x.

Turns out that an optimization that mixed in a role into the given
iterator (in case of a PredictiveIterator) somehow had *very* adverse
efficiency effects.  The goal of the optimization was to provide a
fast way to tell how many permutations could be produced.  This *was*
indeed a lot faster, but at the expense of some very serious performance
issues in normal usage.

Spotted by Jan-Olof Hendig++

Removed the iterator mixin code, as there were no clients of it anymore.

This breaks one spectest of the combinations iterator: this has now
also gained PredictiveIterator semantics, but the test assumes it does
not.  Soon to be fixed.