Refactor `Rakudo home` and `NQP home` handling
It is now not set as an HLL symbol anymore, but instead can be retrieved
via a method on the Compiler object. This deduplicates the logic previously
present in both, src/main.nqp and src/rakudo-debug.nqp.

Also NQP now determines and provides the `NQP home` itself in its Compiler
object so we don't need to redetermine it in rakudo.