Make X::Method::NotFound exception more user-friendly
The exception can now provide user with a few additional tips depending
on the context of the method call and what kind of mistake the user
possibly made. Tips include, in addition to the traditional suggestion
list of like candidates, the following two:

- `self!"private"()` syntax if `self."!private"()` form is presumably
  used.

- If a private method invoked but all suggested candidates are public
  then tip the user about possibly using public method call. Same advise
  is given for public invocation if all suggestions are private.

No private methods are suggested if call is made outside of a class
body. The condition is checked by matching invocator class to `$?CLASS`.

Measures are taken not to flood the user with surplus tips.