Add a command line option `--rakudo-home` use it during build
The precedence order of the different ways to set the Rakudo home is now:

1. Commandline option `--rakudo-home`
2. Environment variable `RAKUDO_HOME`
3. Static rakudo home set by `Configure.pl` when building a non-relocatable
   build
4. Dynamically determined by looking at the executables path - that's the
   relocatable build

During the build the first rakudo that is run (`RUN_RAKUDO` Makefile
variable) now passes the `--rakudo-home` parameter and sets the Rakudo
home to `$BUILD_DIR/gen/build_rakudo_home`. This makes sure this first
rakudo used during build does not use any files found in the install
directory.

This fixes #3798.