MoarVM's struct MVMCollectable's `flags` is now `flags1` and `flags2`
As part of the fix for the data race described in MoarVM/moarvm#1336
the flags member of struct MVMCollectable needs to be changed from a
single MVMuint16 to two MVMuint8 values. As `p6setfirstflag` and
`p6takefirstflag` directly manipulate this struct, they need to be
updated to reflect the changes.

Implemented with conditional compilation so that it will work with MoarVM
before and after these changes. Once MoarVM is updated, the conditional
compilation can be removed.