Make error when sorting with junctions more awesome
You cannot sort a list that contains Junction values.  The reason for
this, is really that we do not have a sensible way of collapsing a
Junction into an Order value.

Instead of dying with an internals error about not being able to
coerce a Junction to an int, it now dies with a X::Cannot::Junction
exception.

Spotted by AlexDaniel++