Make X::Method::NotFound case-insensitive
When an X::Method::NotFound exception is thrown, it's .message method
suggests up to four similar methods that the user might have meant.
However, it computes "similar" methods based on a case-sensitive
comparison of the user-supplied method and methods available on the
object.

This case sensitivity can result in methods not being
suggested even though they are largely similar (or even identical) to
the method the user supplied, apart from case. This can be a
particular problem for methods that new users may have forgotten are
all uppercase.  For example, `why` and `WHY` share no letters when
compared in a case-sensitive way.

This commit resolves this issue by calling `.fc` before performing the
comparison.  This may *very* slightly slow the comparison, but I am
not concerned – when generating the message from X::Method::NotFound,
the program is very likely to be in an unrecoverable state, where
extremely slight performance issues are unlikely to have an impact.

This commit updates the corresponding test, and hopefully makes it
slightly less fragile – it no longer tests against the suggested
methods being an exact match, but just tests that private methods are
correctly excluded.