Allow for more parallelism when precompiling modules
Instead of locking the whole precomp store whenever we precompile a
module or re-check dependencies, we now use one lock file per precomp file.
With this multiple processes may precompile different files and store
them in the same precomp store concurrently.

CompUnit::PrecompilationRepository.precompile not only writes the precomp file
but also an accompanying .repo-id file. This is why we still use the same lock
for the precomp file itself and the .repo-id file (instead of individual locks
for these files).