Handle the case of cmp returning a Junction
As AlexDaniel++ pointed out, we should *also* check whether cmp
returned a Junction, which may also happen in some situations.

This does *not* fix the case of sorting lists that have Junctions
in them.  The case of sorting Maps/Hashes is even more interesting:
it stringifies the Map/Hash for comparison, and this uses sort, and
dies doing *that*.

Also abstract the error throwing into a separate sub: DRY!