Globalize local compare function
CMP_DISALLOW_JUNCTIONS is now a global "implementation-detail" sub
that provides a faster infix:<cmp> that does not allow Junctions and
returns a native int rather than an Order enum.  This is now also used
by the @a cmp @b logic.

Improve the logic a bit more, so that if infix:<cmp> returns a Junction
inside CMP_DISALLOW_JUNCTIONS, it will also throw.

This does *not* fix sorting a list of Maps/Hashes.  The reason for this
is that there is no specific candidate for %a cmp %b, so the generic one
is used, which stringifies its parameters.  And stringification of a Map/Hash
uses .sort, so that problem is more general.