Externalize suggestions and tips
The X::Method::NotFound error is hard to test for.  This commit
adds the "suggestions" and "tips" method, so that one can test for
(not) expected suggestions, and whether any tips were generated.

Also only use numeric weights in suggestions: there was a mix of
strings and integers, that can't be right  :-)