Introducing the allomorph class
- a subclass of Str
- IntStr/NumStr/RatStr/ComplexStr are now subclasses of allomorph
- replaced several nqp::ifs by ternaries
- streamlined many coercions, specifically RatStr.Numeric
- simplified 16 infix:<eqv> candidates into a single one
- need further research to do same for infix:<cmp> and infix:<===>
- make test and make spectest clean

The reason for this is to make it easier and more extendable to handle
allomorphs.  Having an allomorph class should allow custom numeric
objects to also include their allomorphic version.  And having an
allomorphic class should make typechecking for allomorphs more centralized
as is proven by the infix:<eqv> simplification.