Make package body be a Block, not Blockoid
Which greatly simplifies matters so far as code-gen goes, but is also
more correct: a package with `unit` scope should not imply a lexical
scope, so RakuAST::Package itself being a RakuAST::LexicalScope would
have caused trouble once we came to support that. This means we spit out
the QAST for methods in the class body.