Provide a valid NQP-home to the build dir rakudo runner
When building a relocatable rakudo, NQP-home is not set to a static value,
but determined dynamically based on the path of the program currently
running. This fails for the build dir runners, as they are not located in
the installation folder. We work around this, by setting the NQP_HOME env
variable to the NQP-home directory in the installation directory.