Remove ssh/rsync bits
Previously, a laptop/desktop setup was used to share tasks
between a powerful desktop for building/testing and a simple laptop
to sign/upload things.

This setup was useful before, but:

- It is not made optional/configurable, so those who don't need it (actually,
  just me) had to either dirty edit akefile or run separate targets manually,
  which can ruin releases (and quite possible did)
- Most machines nowadays can do a release just fine (arguable). A 2016 laptop
  with 2 cores could do it and release managers probably use something more
  interesting.

If anyone considers this setup to be worthy, please re-introduce it,
making it completely optional and turned off by default or
become a release manager while also reverting this commit.