Fix occasional test failures caused by overly clever reproducible-builds.t
reproducible-builds.t assumed that a whole precomp store is locked for writing
and therefore it would suffice to lock for an arbitrary precomp id. This
assumption is no longer true and the trick it played actually violates a
different assumption: that we only lock one precomp file at a time.

Better to stop playing tricks and just copy the test files to a private
repository where other tests cannot interfere (which is what locking the
store for the duration of the test tried to fix).