Revise deprecation warnings to show only on v6.d
The previous version of this change inadvertently displayed the
deprecation warning on v6.c and later; this fixes the code to only
display the warning on v6.d.

I would have liked to put both the sub `parse-names` and the method
`parse-names` in the new `core.d/Str.pm6` file for clearer
organization.  However, I could not see a way to correctly modify a
`Str` method from inside that file.  If someone can point me in the
direction of how to do that in a clean way, I'd be happy to revise
this change.