Preparations for BEGIN-time code evaluation
With RakuAST we'll:

* Interpret simple things
* Send the rest through the compilation pipeline

The current Rakudo compiler will "interpret" constants, but we'll go a
bit futher with RakuAST-based compilation. (This will also mean we can
get a bit further without wiring up the compilation-based approach.)
Given many things we need to evaluate at BEGIN time are quite simple,
this should hopefully mean we'll have better compilation performance.
(The big target - though we're certainly not there yet - is to do
this on role bodies, which could greatly reduce the cost of compiling
code containing roles.)

This adds support for `BEGIN 42` (and other literals), as a proof of
concept of the idea, but it's probably also enough for `use lib <lib>`
to be wired up too.