Introducing Formatter class and associated Grammer
This is basically taking my work on the "newer-sprintf" branch into
the RakuAST era, and also making that work a first class citizen.

The Formatter class takes a sprintf format string and returns a
Callable that can be called, with the necessary parameters, to have
those parameters formatted into a string.

So basically:

    say sprintf "%02d: %s", 42, "foo";

can also be expressed as:

    say Formatter.new("%02d: %s")(42, "foo")

But you can go further than that:

    my constant &lister = Formatter.new("%02d: %s");

    say lister(42,"foo")

would move the creation of the Callable into compile time, with
the added benefit that any errors in the format string would be
caught at compile time.  Various types of syntactic sugar for this
are being mulled at the moment, with q:fmt/%02d: %s/ being one of
them.

This commit still just creates strings for an EVAL to handle, so
no RakuAST specific bits yet.  This will be attempted in the near
future and will be an indication as to how far the RakuAST branch
has gotten.