BUILD: Fix RUN_RAKUDO to use the dummy Rakudo home
The command to execute Rakudo during the early stages of the build process
needs to have a Rakudo home set to an empty dummy directory to make sure it
does not try to access the installation directory. If it would, it would
create directories out side of the build directory during the compilation
phase and result in failures if a previous installation is already located
in the install dir. The `--rakudo-home` Configure parameter doesn't exist
(any more?). Replace it with a perl one-liner that sets the Rakudo home via
an environment variable.