De-huffmanize Capture.raku
The \(...) syntax is confusing to a lot of newbies, especially coming
from Perl.  In order to de-emphasize that syntax, change the .raku
output to use the .Capture coercion syntax, which I think makes it
much more clear what is meant.

This will require *one* change in spectest, that specifically tests
for the \(...) format in output:

    t/spec/S12-introspection/walk.t   test 23

Deprecating / removing the \(...) syntax would be something for
multiple language versions.