Fix regression in Test::Fuzz
For some reason still unclear to me, Test::Fuzz's internals are affected
by df5b3d5308bd175739a6 .  For some reason, this appears to be related
to the use of a TWEAK method in the BUILDPLAN of X::Method::NotFound.
So instead of using TWEAK, put the code of TWEAK into a private method
and call that as appropriate, to get the same result.

Note therefore that this is *really* a workaround.  Further research
would be needed to really get to the bottom of this.