Add deparse support for QuotedWords
This turned out to be quite a bit more involved than expected, because
fixing tests in one place, would break tests in another.  In the end it
was decided that probably there's an error in the codegen of the
QuotedWords class.  To be more precise:

- having processors "words", "val" should act as < >
- having processors "quotewords", "val" should act as << >> (NYI)

The tests have been adapted accordingly, which means that one test in
strings.t now breaks for the AST version.