sub MAIN: Negation and explicit values with bundling and single parameter
When using the %*SUB-MAIN-OPTS<bundling> option negation and explicit
values are ambiguous when actually bundling multiple params together.

`-abc=foo` # a = True, b = True, c = "foo"? a = b = c = "foo"?

When only a single param is given though, there is not ambiguity.

`-a=foo` # a = "foo"

This change makes the latter possible.