Make some set ops prec 'non' (structural)
This fixes the problem described in R#3945 and is spectest clean.

The problem was caused by Z(&) basically codegenning to:

   infix:<(&)>( (left,right) )

for each entry on either side of the meta-op, instead of:

   infix:<(&)>( left, right )

The former just creating a Set of all elements in left/right (because
it was using the single arg candidate), rather than actually applying
the op to the left/right side.

This most likely also fixes the problem for other meta-ops, such as X.