Use additive / multiplicative precedence
As AlexDaniel++ pointed out, the precedence setting in the grammar
was forgotten to be updated as well.  Changing that accordingly,
brought up spectest failures.

Change the precedence to "additive" or "multiplicative" (for (.) ).
This also fixes the issue that was at the base of this PR, and feels
more natural for the type of operator we're talking about here, as
you could see each of these operators as a type of addition (or
multiplication).