Add marker to set ops, adapt metaops for it
This introduces two new precedence hashes: "set_junctive_and" and
"set_junctive_or" that are identical to "junctive_and" and "junctive_or"
except that they have a true "set" key in the hash.

Adapt the affected set operators to use these precedence hashes
(keeping the same precedence level as before).

Since these are only needed in Rakudo *after* the setting has been
built, there is no need to add these to the precedence hashes in
Grammar.nqp.

Finally, add a check for the "set" key in the precedence hash and
act accordingly.